PCA Modifications (#383)
* n_samples < n_features case in PCA enabled

* PCA patching optimization

* dtype check removing

* _fit() patching
