+++ daal4py/sklearn/decomposition/__init__.py
@@ -1,3 +1,3 @@
-from .pca import PCA
+from ._pca import PCA
 
 __all__ = ['PCA']
