--- daal4py/sklearn/decomposition/pca.py
@@ -1,26 +1,0 @@
-#
-#*******************************************************************************
-# Copyright 2014-2020 Intel Corporation
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#******************************************************************************/
-
-from sklearn import __version__ as sklearn_version
-from distutils.version import LooseVersion
-
-if LooseVersion(sklearn_version) >= LooseVersion("0.23"):
-    from ._pca_0_23 import *
-elif LooseVersion(sklearn_version) >= LooseVersion("0.22"):
-    from ._pca_0_22 import *
-else:
-    from ._pca_0_21 import *
